read all comment limiting

+++ backend/crud/comment.py
@@ -24,15 +24,15 @@ class CommentCRUD:
             raise e
 
     @staticmethod
-    def read_all_comment(db: Session) -> List[Comment]:
+    def read_all_comment(db: Session, skip: int = 0, limit: int = 10) -> List[Comment]:
         try:
-            comments = db.query(Comment).all()
+            comments = db.query(Comment).offset(skip).limit(limit).all()
             return comments
         except Exception as e:
             raise e
 
     @staticmethod
-    def read_comment_by_commenter(commenter_id: str, db: Session):
+    def read_comment_by_commenter(commenter_id: str, db: Session) -> List[Comment]:
         try:
             comments = (
                 db.query(Comment)
@@ -45,11 +45,18 @@ class CommentCRUD:
             raise e
 
     @staticmethod
-    def delete_comment_by_id(comment_id: int, db: Session):
+    def delete_comment_by_id(comment_id: int, current_user: str, db: Session) -> Comment:
         try:
             comment = db.query(Comment).filter(Comment.comment_id == comment_id).first()
+
+            # the comment doesn't exist
             if comment is None:
                 raise Exception(f"comment ID {comment_id} doesn't exist")
+
+            # the comment exists but the user is not the owner
+            if comment.commenter_id != current_user:
+                raise Exception('You can NOT delete it!')
+
             db.delete(comment)
             db.commit()
             return comment
